requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.3.0
    # via passlib
cffi==1.17.1
    # via cryptography
click==8.1.8
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via click
cryptography==44.0.2
    # via python-jose
databases==0.9.0
    # via -r requirements.in
dnspython==2.7.0
    # via email-validator
dotenv==0.9.9
    # via -r requirements.in
dynaconf==3.2.10
    # via -r requirements.in
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.115.12
    # via -r requirements.in
greenlet==3.2.0
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
jinja2==3.1.3
    # via -r requirements.in
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
passlib[bcrypt]==1.7.4
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.6.4
    # via
    #   -r requirements.in
    #   fastapi
    #   sqlmodel
pydantic-core==2.16.3
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via dotenv
python-jose[cryptography]==3.4.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
rich==14.0.0
    # via typer
rsa==4.9.1
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.29
    # via
    #   -r requirements.in
    #   alembic
    #   databases
    #   sqlmodel
sqlmodel==0.0.24
    # via -r requirements.in
starlette==0.46.2
    # via fastapi
typer==0.15.2
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
uvicorn==0.29.0
    # via -r requirements.in
